выполняет домашнее задание 5.15
